Implement display_protected_tag_from_post_list settings Cleanup index.js for forum with extended functions files to separate extended or override sections.
